+++ darts/timeseries.py
@@ -3238,7 +3238,7 @@ class TimeSeries:
         # TODO: check
         if method == "pad":
             new_xa = resample.pad()
-        elif method == "bfill":
+        elif method in ["bfill", "backfill"]:
             new_xa = resample.backfill()
         else:
             raise_log(ValueError(f"Unknown method: {method}"), logger)
@@ -3360,6 +3360,7 @@ class TimeSeries:
         forecasting_safe: Optional[bool] = True,
         keep_non_transformed: Optional[bool] = False,
         include_current: Optional[bool] = True,
+        keep_names: Optional[bool] = False,
     ) -> Self:
         """
         Applies a moving/rolling, expanding or exponentially weighted window transformation over this ``TimeSeries``.
@@ -3458,10 +3459,14 @@ class TimeSeries:
 
         keep_non_transformed
             ``False`` to return the transformed components only, ``True`` to return all original components along
-            the transformed ones. Default is ``False``.
+            the transformed ones. Default is ``False``. If the series has a hierarchy, must be set to ``False``.
 
         include_current
             ``True`` to include the current time step in the window, ``False`` to exclude it. Default is ``True``.
+
+        keep_names
+            Whether the transformed components should keep the original component names or. Must be set to ``False``
+            if `keep_non_transformed = True` or the number of transformation is greater than 1.
 
         Returns
         -------
@@ -3611,6 +3616,53 @@ class TimeSeries:
         if isinstance(transforms, dict):
             transforms = [transforms]
 
+        # check if some transformations are applied to the same components
+        overlapping_transforms = False
+        transformed_components = set()
+        for tr in transforms:
+            if not isinstance(tr, dict):
+                raise_log(
+                    ValueError("Every entry in `transforms` must be a dictionary"),
+                    logger,
+                )
+            tr_comps = set(tr["components"] if "components" in tr else self.components)
+            if len(transformed_components.intersection(tr_comps)) > 0:
+                overlapping_transforms = True
+            transformed_components = transformed_components.union(tr_comps)
+
+        if keep_names and overlapping_transforms:
+            raise_log(
+                ValueError(
+                    "Cannot keep the original component names as some transforms are overlapping "
+                    "(applied to the same components). Set `keep_names` to `False`."
+                ),
+                logger,
+            )
+
+        # actually, this could be allowed to allow transformation "in place"?
+        # keep_non_transformed can be changed to False/ignored if the transforms are not partial
+        if keep_names and keep_non_transformed:
+            raise_log(
+                ValueError(
+                    "`keep_names = True` and `keep_non_transformed = True` cannot be used together."
+                ),
+                logger,
+            )
+
+        partial_transforms = transformed_components != set(self.components)
+        new_hierarchy = None
+        convert_hierarchy = False
+        comp_names_map = dict()
+        if self.hierarchy:
+            # the partial_transform covers for scenario keep_non_transformed = True
+            if len(transforms) > 1 or partial_transforms:
+                logger.warning(
+                    "The hierarchy cannot be retained, either because there is more than one transform or "
+                    "because the transform is not applied to all the components of the series."
+                )
+            else:
+                convert_hierarchy = True
+
         raise_if_not(
             all([isinstance(tr, dict) for tr in transforms]),
             "`transforms` must be a non-empty dictionary or a non-empty list of dictionaries.",
@@ -3688,9 +3740,22 @@ class TimeSeries:
                 f"{'_'+str(min_periods) if min_periods>1 else ''}"
             )
 
-            new_columns.extend(
-                [f"{name_prefix}_{comp_name}" for comp_name in comps_to_transform]
-            )
+            if keep_names:
+                new_columns.extend(comps_to_transform)
+            else:
+                names_w_prefix = [
+                    f"{name_prefix}_{comp_name}" for comp_name in comps_to_transform
+                ]
+                new_columns.extend(names_w_prefix)
+                if convert_hierarchy:
+                    comp_names_map.update(
+                        {
+                            c_name: new_c_name
+                            for c_name, new_c_name in zip(
+                                comps_to_transform, names_w_prefix
+                            )
+                        }
+                    )
 
             # track how many NaN rows are added by each transformation on each transformed column
             # NaNs would appear only if user changes "min_periods" to else than 1, if not,
@@ -3745,6 +3810,15 @@ class TimeSeries:
         # revert dataframe to TimeSeries
         new_index = original_index.__class__(resulting_transformations.index)
 
+        if convert_hierarchy:
+            if keep_names:
+                new_hierarchy = self.hierarchy
+            else:
+                new_hierarchy = {
+                    comp_names_map[k]: [comp_names_map[old_name] for old_name in v]
+                    for k, v in self.hierarchy.items()
+                }
+
         transformed_time_series = TimeSeries.from_times_and_values(
             times=new_index,
             values=resulting_transformations.values.reshape(
@@ -3752,7 +3826,7 @@ class TimeSeries:
             ),
             columns=new_columns,
             static_covariates=self.static_covariates,
-            hierarchy=self.hierarchy,
+            hierarchy=new_hierarchy,
         )
 
         return transformed_time_series